do not double-define pytest fixture (mypy issue)
